+++ pennylane_calculquebec/processing/steps/__init__.py
@@ -12,3 +12,12 @@ from .decompose_readout import DecomposeReadout
 from .gate_noise_simulation import GateNoiseSimulation
 from .readout_noise_simulation import ReadoutNoiseSimulation
 from .print_steps import PrintResults, PrintTape, PrintWires
+
+from pennylane_calculquebec.processing.processing_error import ProcessingError
+class StepsError(ProcessingError):
+    """Error related to steps."""
+    def __init__(self, message: str):
+        predefined = "Steps Error/"
+        full_message = f"{predefined}{message}"
+        super().__init__(full_message)
+        self.message = full_message
